# Use go build -i for faster proxy builds