This will be redundant in go 1.10, but it seems to speed up larger test suites by about 10x.