@lox 2018-01-26
Or maybe it doesn't, see below benchmark 🤔

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkCreatingProxies-4     519177956     632174147     +21.76%
BenchmarkCallingProxies-4      11769882      12569427      +6.79%

benchmark                     old MB/s     new MB/s     speedup
BenchmarkCallingProxies-4     5.10         4.77         0.94x

benchmark                      old allocs     new allocs     delta
BenchmarkCreatingProxies-4     265            262            -1.13%
BenchmarkCallingProxies-4      342            342            +0.00%

benchmark                      old bytes     new bytes     delta
BenchmarkCreatingProxies-4     87184         86712         -0.54%
BenchmarkCallingProxies-4      132279        132367        +0.07%
```